+++ pdm/cli/commands/init.py
@@ -63,7 +63,7 @@ class Command(BaseCommand):
                         project.python = PythonInfo.from_path(get_venv_python(path))
                     except Exception as e:  # pragma: no cover
                         project.core.ui.echo(
-                            f"Error occured when creating virtualenv: {e}\n"
+                            f"Error occurred when creating virtualenv: {e}\n"
                             "Please fix it and create later.",
                             style="red",
                             err=True,